Ingestion layer for ingress/routes
Adds ingestion layer for ingresses and routes. The ingestion layer adds the
ingresses and routes to the work queue if and only if the status fields in
these ingresses are populated with the load balancer IP addresses. Otherwise,
it just rejects them.
Test cases added:
- Verify ingress/route addition and key added
- Verify ingress/route update (backend service) and key added
- Verify ingress/route rejection (no IP in the status field)
- Verify ingress/route update (IP added now) and key added

Note that the ingresses are in this commit considered for extensions/v1beta1,
for later versions of ingress, we will add them later.