Graph layer for http routes
Added layer 2 (graph) for routes ingestion. We consume the keys fetched
from the queue (filled up by the ingestion layer). For each of these
objects, we figure out what operation was performed on this key. Based on
that, if its an add operation for a route, we create a new GSLB service
graph object. The name of a GSLB service is the hostname of the route, for
now. We also add a member to this GSLB service which is the IP address, again
fetched from the route.

For an update operation, the GSLB service must exist. So, we just go ahead
and add a member to the GSLB node object. For a delete operation on a route
object, we go ahead and delete the GSLB graph for that GS name.

In order to determine the GS Name for a route object, we maintain a map
of route to hostname. The reason being, for a route delete operation, we don't
have the route data, as it is already deleted from the existing accepted and
rejected route caches. The map is updated whenever there's an add or update
operation.

After we create/update/delete a GS graph, we add its key to the GraphLayer
queue, for which the upcoming rest layer is the consumer.