AV-84213,AV-82786 #comment GDP selection unit tests, GDP status updates
This commit adds unit tests for GDP selection and match rules checks. More
intricate checks will be subsequently added. These are the checks which this
PR covers:
- Create: Select all objects from one cluster.
- Create: Select few objects from one cluster.
- Create: Select all objects from all clusters.
- Create: GDP object with typo/wrong object names.
- Create: Multiple GDP objects on the same namespace (multi-cluster).
- Update: Reject all -> Select few (multi-cluster).
- Update: Empty match clusters -> One match cluster. (multi-cluster config).
- Update: Switch clusters (from cluster1 to cluster2).
- Update: Typo/Wrong cluster name in the cluster contexts.
- Create: No cluster selected.
- Create: No objects selected.

This patch also covers updating the status field of the GDP CRD based on
the GDP's semantic checks. Once a GDP object is created, we first check
the sanity of that object wrt the object types in the matchRules, the
operations, missing key/value in label, wrong cluster context(s), wrong
traffic split values (negative values/ not in the range of 1-20) etc. If
everything's fine, "success" message will be populated. These checks were
required from the unit tests pov, as there was some way needed to figure
out the problems in the GDP object.

The status field will now be populated with this value:
```
status:
  errorStatus: ""
```