AV-84213: Check if a leader is not a leader anymore and accordingly stop syncing rest operations
If an error is encountered during a rest operation to the avi controller,
we check the status codes of this error. If its a 400 and if the error
message contains that "Config Operations can ONLY be done on leader", we
mark the controller to be non-leader and return from there. This means,
we don't update any cache in layer 3 and syncing is stopped from layer 3
to the controller.

During full sync, we check whether this controller is the GSLB leader or
not, and based on that we mark the controller to be a leader. After which,
any object updates WILL be synced. This is done to ensure that if a site
becomes a leader again, the full sync should detect it and normal operations
should resume.