AV-85020 #comment Changed the filter structure, accept/reject rules
According to the new design of the GDP CRD, changed the filters. Also
added a namespace informer for member clusters. For a change in the
GDP object, we still re-evaluate the accepted/rejected objects but the
rule structure has been changed. For a namespace added, we check the
GDP object for namespace selector. If there is one, we check this
namespace's labels against this namespace selector and add it to the
filter as well. For a namespace change in the filter, we again, re-
evaluate the ingress/service objects.