AV-84460 #comment GSLBConfig object status update
This PR adds support for checking the current status of the GSLBConfig
object. When a new GSLBConfig object is added, we do the following:
1. check whether a GSLBConfig object was already accepted, if
   yes, we update the new GSLBConfig object's status field saying, a
   GSLBConfig object already exists, can't accept another one.
2. check the validity of this object, and assign this an invalid
   config error if the config is invalid.
3. check the connectivity with the controller, if it failed, update
   the status with connectivity failure error.
4. check the creation of the member clusters kubeconfig file and
   update the status accordingly.
5. update the status to boot up syncing, and start syncing objects
   for full sync.
6. if all the above passed, we update the status saying-
   "success: accepted GSLBConfig"

If AMKO receives a GSLBConfig object update, we do the following:
1. check the accepted gslbconfig object's name and namespace with
   the new object's name and namespace.
2. if the above check passed, update status saying gslbconfig object
   edited, a restart is required.

If AMKO receives a GSLBConfig object delete, we check whether this is
the same GSLBConfig object that's already accepted, if yes, we mark
the accepted flag as false and be ready for a new GSLBConfig object.