AV-86528 #comment Fix intermittent non-selection issue with ingresses/services
If a label is deleted for a namespace/service/ingress while amko is down,
then in the next reboot, amko as it was, wasn't taking into account
the non-selection of those objects. Which means that, if a GS was
previously created because of these objects, and now we don't select
those objects after a reboot, we weren't deleting the GS(s).

This patch first creates the current avi cache, builds a list of all
GS graphs based on selected objects and then, finds out the diff between
the current GS graphs and the avi cache and publishes those keys to
layer 3. We also create a dummy GS graph for these cases because layer
3 (rest layer) will only delete GS objects for whom the members length
is 0.

An additional change fixes a typo in the GDP config schema.