AV-90719 #comment update the labels in the rejected stores for objects
We aren't updating the labels if a k8s/openshift object is getting
rejected in the update event handler. The labels for these objects
in the rejected stores contain stale values. We need to store the
new objects in these stores, instead of moving them from accepted
to rejected stores.

For namespace update, we already do this instead of moving the
objects.