AV-92229 #comment Check whether controller is a leader during full sync
If a controller gets changed to a follower site while GSs are getting
synced, we set the controller to follower site in AMKO and stop
syncing from layer 3 today. However, AMKO doesn't revert the condition
and resume syncing if the controller is set to leader again.

Here, we check the controller status on every full sync. If the controller
becomes a follower while syncing the GSs, we stop syncing from layer 2
itself. But we still continue to listen to the kubernetes/openshift
objects and continue to build our GS graphs. On every full sync, we
will now check if the controller has become the leader. If it has become
the leader, we publish all the keys from graph layer to the rest layer.