AV-92417,AV-91727 #comment add system-gslb-tcp as the healthmonitor for GS objects
AV-92417:
Today we use System-GSLB-Ping as the default health monitor for all the
GSLB services. However, if a vip goes down because a cluster goes down,
the VIP is still reachable via ping. So, instead, we will use the
System-GSLB-TCP healthmonitor for routes and ingresses. For services,
we check for any of the port protocol, if it is TCP, we use the TCP
health monitor, if it is UDP, we use the UDP health monitor.

During bootup full sync, we fetch all the GSLB health monitors and add
them to the cache. We also cache them by UUID, required when parsing the
GSLB services individually. This we need to do because in layer 2, we
don't have a UUID for a non-existent health monitor, and hence, we have
to use the health monitor name. In layer 3 too, we use the HM name for
consistent checksum calculation. If we fail to delete the HM due to any
controller error, we publish the key to the retry layer.

AV-91727:
We check the routerName field in the route's status message to make
sure that the IP address that we get is from AKO.

Fixes:
- Full sync wasn't taking the openshift routes into account.
- Couple of logs fixes.