AV-94412 #comment Fix to not delete the model key from layer 3
If layer 3 is in the process of deleting the previous GS object, and amko
gets another create for the same object in layer 2, layer 2 creates the
model for the GS. However, as soon as layer 3 is done deleting the object
from AVI, it deletes the key from the model cache. This happens in case
of rapid create/delete of the GDP objects.

Fix is to not delete the model key from layer 2 cache after the object
deletion in layer 3. We are anyway taking care of deleting the members
from the model cache.