AV-94414, AV-94409, AV-94163 #comment Added wrapper for rest calls, introduced layer 2 delete cache
Additional changes include:
- Imported new SDK changes (with controllerStatusCheck flag)
- Imported new utils package (disabling the controller status check while
  initializing the avi clients)
- Added a few more debug logs

Wrapper for rest calls has a timeout of 600 secs. If a request to the Avi
controller takes more than 600 secs, we timeout in AMKO's rest layer and
put that key in the slow retry queue. If we get any other EOF/connection
failures, we put the key in the slow retry queue. Note that we don't
use avi sdk's CheckControllerStatus() function any more, so there will be
no more retries from avi session in case of any connection failures.

The delete GS logic has been changed. We introduce a new delete cache which
gets initialized in layer 2. If we need to delete a GS based on 0 members
in layer 2, we delete the key from gs model cache and put the key and the
model in the delete cache. Once the rest layer recieves the key, it checks
if the key only exists in the delete cache and deletes the object and the
associated health monitor (if any). It may happen that, while that just
before rest layer for a key starts, the graph layer can add the key to
the model cache (because it got a new create operation). In that event, we
check if the key exists in the model cache, if yes, we ignore the delete
cache.