AV-94507 #comment Fix UDP health monitor for Gslb services
UDP health monitor depends on the service type that its actually monitoring.
So, the health monitor is application specific. However, by default, the UDP
health monitor just sends ICMP requests to the backend servers. For a default
UDP health monitor to be created, the udpRequest can be any random string. If
a user then wants to change it to their application specific request string,
they can just change the health monitor object. The port however, can't be
changed and it depends on the LB service for whom the health monitor got created.

This commit also includes some log fixes.