Fix the host fqdn check in route status field
If we change the route's host in spec to an invalid hostname (not supported
by avi dns), amko takes that host fqdn as a valid host fqdn. We will now
check if the spec's hostname matches the status hostname field.

Also adds the hostname population in the route status field.