Skip initializing member cluster if HostRule CRD absent
If HostRule CRD is not installed in a member kubernetes/openshift
cluster, AMKO will skip the initialization of the member cluster and an
error will be logged.

This also fixes the number of minimum hm up for hm refs. This would
mean, if a user has specified custom health monitor refs, all the health
monitors have to be UP for a GSLB pool member to be up.