Add GDP invalid Health Monitor transition cases
Added a test case where initially a valid GDP object is added, then we
add a list of valid and invalid health monitor refs in the GDP object.
We verify that the members are unchanged and the GDP object was
rejected. We then fix the GDP object by removing the invalid health
monitor ref and adding a valid health monitor ref.

The invalid health monitor here means that the health monitor is not
federated which should be rejected by AMKO.