VS UUID fix and aviserver integration test cases
Bugfix:
If annotations are removed for an object in the VS UUID mode, we ignore
any updates to that member. This is incorrect, as we should be deleting
the member on removal of annotations. The fix is in layer 2, where,
upon seeing that the annotations are absent, we delete the gs member
from the gs graph and send the key to layer 3.

Integration tests enhancements:
- Added middlewares for Post/Put calls for GS and HM objects.
- Added library handlers to insert the middlewares.