Fix custom HM delete issue while bootup
Bug: AMKO tries to delete the custom health monitor while booting up.

Conditions:
- Pre-existing Gslb Services on the controller.
- No ingress/routes or no GDP object.
- AMKO is booting up.

Cause:
While booting up, AMKO fetches the list of existing GSs. It then tries
to fetch the ingresses from the member clusters. If AMKO can't find
an ingress/route for an existing GS, it triggers a delete of that
GS object and it's corresponding Health monitor. In the rest layer,
AMKO checks if the GS graph has custom Health Monitors, if yes, it won't
delete the HMs.
However, during bootup, we don't have a GS graph, as the corresponding
cluster objects/GDP object is removed. Hence, the rest layer check of
custom HM won't work.

Fix:
For each health monitor that we try to delete in the rest layer, we
need to check if the health monitor was created by AMKO by checking the
prefix of the health monitor name.