Fix GSLBHostRule accepted to rejected state
If a GSLBHostRule was in accepted state, and if an invalid field
was added as part of this object, AMKO marks it as rejected. However,
it doesn't reset the GS object to the GDP properties. The fix involves
a change in the Update logic, where, if an object moves from accepted
to rejected, we delete the GS rules from cache and add a delete key
for the previously accepted gs fqdn.