Add infinite retry for errors other than 404
While fetching objects to verify refs for GDP object, we don't retry
infinitely today. This may pose a problem in situations where the
connection between AMKO and the Avi Controller has broken for sometime.
To handle this, we can retry infinitely for get requests for the refs
present in the GDP object.