Fix checksum calculation for GSs
There are a couple of bugs today in checksum calculation. First one,
if there is a GDP object, and there's a GSLBHosRule object, and since,
we don't fetch the GSLBHosRule objects during boot up, we end up
making an unnecessary PUT request. Fix is to fetch the GSLBHosRule
objects during boot up.

Second one, if the health monitor used is `System-GSLB-Ping` or any
other health monitors which doesn't have the `monitor_port` field set,
we end up not storing that health monitor in our cache and hence, we
end up making PUT request on every reboot.