Add default algorithm in checksum
If pool algorithm settings are not present in the GDP object, we don't
consider the default algorithm in the checksum calculation, although,
we add it as part of the GSLB Service in rest layer. This commit adds
the default pool algorithm as part of the checksum calculation.

Also, we always add TTL field to the GSLB Service, even when it isn't
a part of the GDP/GSLBHostRule object. This commit fixes that by
checking the TTL value first and then adding it as part of the
GSLBService in the rest layer.