Optimize health monitor and site persistence lookup
During bootup today, AMKO tries to fetch the referred objects like
Health Monitor refs and Site Persistence Profile refs in `GDP` and
`GSLBHostRule` objects. However, repeated GET requests for these
objects is not desirable.

This fix creates a site persistence cache during bootup and leverages
the existing health monitor cache for such lookups. If the referred
object is not found, we mark the object as invalid (as before).
Non-bootup operations like `UPDATE` for `GDP` and `ADD/UPDATE` for
`GSLBHostRule` will still follow the pre-existing logic of fetching
the refs from the controller, so as to avoid stale entries.