Add support for federating GSLBConfig and GDP objects by AMKO
Currently, AMKO doesn't support running in federation mode, where it
can federate the required configurations across member clusters. This
requires either a 3rd party backup and restore solution for users or
they have to manually copy and restore the AMKO configs.

This PR adds support for a federator which automates the federation of
`GSLBConfig` and `GDP` objects across all member clusters whenever they
are changed. AMKO will now run in either a leader mode or a follower
mode. The leader AMKO is responsible for federation and for syncing objects
to the GSLB leader Avi Controller. Passive/Follower AMKOs run in standby
mode.

A new CR called `AMKOCluster` is introduced where users would have to
input the clusters participating in federation and specify whether AMKO
is running in leader/follower mode:
```
  spec:
    clusterContext: cluster1  # current cluster context
    clusters:   # particpating in federation
    - k8s
    - oshift
    isLeader: true    # indicates leader/follower
    version: 1.4.2    # AMKO version
  status:          # current state of federation
    conditions:
    - status: Federation successful
      type: Federation
```

Once the user deploys AMKO, the federator runs its validations and
starts federating if it finds no errors.

Error conditions covered in this PR (where federation won't happen,
and is indicated in the status field):
- Multiple GSLBConfig objects
- Multiple GDP objects
- Multiple AMKOCluster objects
- AMKO is a follower (not a error condition, but is indicated by the
  status)
- Unable to initialize config for any member cluster
- Unable to connect to a member cluster
- Version mismatch with any of the member cluster's AMKO
- Current cluster context not set
- Member cluster context not present in the kubeconfig secret

This PR makes use of the controller runtime for AMKOCluster CR and
uses the controller to also watch on `GSLBConfig` and `GDP` objects.