AMKO bootup checks to ensure that the current instance is the only leader
With introduction of AMKOCluster CRD, we need to check if isleader flag is set to true
when AMKO boots up. If the flag is set to true AMKO would check if any other cluster
is also running as leader, else proceed with normal operation.
If isleader flag is set to false, AMKO would not process any kubernetes objects except
listening to AMKOCluster object.

Whenever isLeader flag isLeader is changed, the API server would be shut down, triggering
a reboot of AMKO.