Fix the conflict with metrics address port
The controller-runtime starts the metrics server in the default
port of 8080 (if a port number is not provided). This has a conflict
with the api server in AMKO. This PR fixes that conflict.

This also adds a fix to send the current AMKO version instead of
the current cluster to the validation function.