Fix parsing error for TTL value for a REST response
This fixes a bug where the checksum doesn't change for the GS cache
update operation, as the response is not parsed properly for the
`ttl` value. Whatever the ttl value, we always set it to `nil` before
calculating the checksum of the GS obtained from the REST response.

This doesn't happen during bootup (as AMKO takes another path), but
during POST/PUT operations for a GS.