Get Site Persistence from cache when GSs are fetched from Avi
Currently, we fetch GSLB Services and parse all the fields. But, the
site persistence references weren't being parsed correctly. Because of
this, the checksum computed was different from layer 2. And this caused
unnecessary updates to the GS. This fix introduces a UUID cache for
the site persistence objects. After fetching the GS object, we parse
the UUID of the referred site persistence object and fetch this object
from the UUID site persistence cache.