Update the health monitor path names only if no references
Currently, when we add a new member in the GS graph pool, we
re-construct the health monitor path list unconditionally. While we do
have a check in layer 3 to make sure that we give priority to the
health monitor references in the `GDP` or `GSLBHostRule` objects, this
can cause unnecessary updates to the GSs in the controller. We
should update the HM paths only if there are no health monitor
references present in the `GDP` or `GSLBHostRule` objects.

Also, parsing the consistent hash mask value wasn't done correctly,
as we were looking for int32. Fixed it to now look for a float64 and
then convert it to int32.