Don't update status if status already reflects that AMKO needs a reboot
Currently, if there's an update to the `GSLBConfig` object, the status
reflects that AMKO needs a reboot since the `GSLBConfig` object got
updated. Meanwhile, if the object had a controller IP which was GSLB
leader previously and got changed to follower, the resync goroutine
determines this and overwrites the status reflecting that the controller
IP provided is not of a leader. This fixes the overwrite issue and
always reflects that AMKO needs a reboot once the `GSLBConfig` object
gets updated.