Add CRDs for MCI and ClusterSet
Implements AV-124910 and AV-124913

These CRDs will be listened to by the Service Discovery Agent and AKO
to create virtual services for applications distributed across multiple
clusters.

MCI contains the definition of an application which is defined as a
collection of endpoints spread across multiple clusters. It contains
the host fqdn and a set of backend services which will be resolved by
a service discovery agent. Definition for this object can be found in
helm/amko/crds/mci_def.yaml.

ClusterSet defines the boundary of multi cluster applications. This
contains a list of clusters from which the MCI endpoints will be
resolved. It has a field for the secret name which would contain the
secret generated from the kubeconfig of the clusters in the clusterset.
Definition for this object can be found in
helm/amko/crds/mci_def.yaml.