Enable multiple AMKOs to co-exist
This PR adds support for adding UUIDs in the created-by field for GSs
and Health Monitors. While booting up, AMKO checks for the UUID field
in the GSLBConfig object. If it finds a valid UUID, this AMKO instance
uses this UUID to fetch/create GSs and HMs. It it can't find an UUID,
this instance then generates a new UUID and adds it to the GSLBConfig
object. This UUID is then used to create/update the GSs and HMs.