Bugfix: Fetch all health monitors including custom HMs
Currently, the commit to support multi AMKO instances introduced a
bug where we weren't fetching health monitors which were created by
users. To support that, we must first fetch all federated health
monitors. Federated health monitors can be of 3 types:
1. HMs created by the current AMKO instance.
2. HMs created by the user (and can refer them in the GDP or GSLBHostRule)
3. HMs created by other AMKO instances.

HMs in category 1 and 2 are the ones that are required to be stored in
the current AMKO instance's cache. Category 3 HMs must be ignored.